replays_load/app/requirements.txt
psycopg2-binary
google-cloud-storage